Upgrade Grav 1.6.31 to 1.7.49.5 and update saturn_custom theme
- Upgraded Grav core from 1.6.31 to 1.7.49.5 via GPM self-upgrade
- Updated saturn_custom theme for Grav 1.7 compatibility:
  - Added deferred asset blocks in base.html.twig
  - Added autoescaping on site.title output
  - Renamed theme class to SaturnCustom for autoloading
  - Bumped blueprints.yaml to 1.7.1 with dependencies
- Fixed duplicate 'markdown' key in user/config/system.yaml

+++ system/src/Grav/Common/Cache.php
@@ -3,19 +3,27 @@
 /**
  * @package    Grav\Common
  *
- * @copyright  Copyright (C) 2015 - 2019 Trilby Media, LLC. All rights reserved.
+ * @copyright  Copyright (c) 2015 - 2025 Trilby Media, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
 
 namespace Grav\Common;
 
+use DirectoryIterator;
 use \Doctrine\Common\Cache as DoctrineCache;
+use Exception;
 use Grav\Common\Config\Config;
 use Grav\Common\Filesystem\Folder;
 use Grav\Common\Scheduler\Scheduler;
+use LogicException;
 use Psr\SimpleCache\CacheInterface;
 use RocketTheme\Toolbox\Event\Event;
-use RocketTheme\Toolbox\Event\EventDispatcher;
+use Symfony\Component\EventDispatcher\EventDispatcher;
+use function dirname;
+use function extension_loaded;
+use function function_exists;
+use function in_array;
+use function is_array;
 
 /**
  * The GravCache object is used throughout Grav to store and retrieve cached data.
@@ -29,42 +37,41 @@ use RocketTheme\Toolbox\Event\EventDispatcher;
  */
 class Cache extends Getters
 {
-    /**
-     * @var string Cache key.
-     */
+    /** @var string Cache key. */
     protected $key;
 
+    /** @var int */
     protected $lifetime;
+
+    /** @var int */
     protected $now;
 
     /** @var Config $config */
     protected $config;
 
-    /**
-     * @var DoctrineCache\CacheProvider
-     */
+    /** @var DoctrineCache\CacheProvider */
     protected $driver;
 
-    /**
-     * @var CacheInterface
-     */
+    /** @var CacheInterface */
     protected $simpleCache;
 
+    /** @var string */
     protected $driver_name;
 
+    /** @var string */
     protected $driver_setting;
 
-    /**
-     * @var bool
-     */
+    /** @var bool */
     protected $enabled;
 
+    /** @var string */
     protected $cache_dir;
 
     protected static $standard_remove = [
         'cache://twig/',
         'cache://doctrine/',
         'cache://compiled/',
+        'cache://clockwork/',
         'cache://validated-',
         'cache://images',
         'asset://',
@@ -74,6 +81,7 @@ class Cache extends Getters
         'cache://twig/',
         'cache://doctrine/',
         'cache://compiled/',
+        'cache://clockwork/',
         'cache://validated-',
         'asset://',
     ];
@@ -115,12 +123,10 @@ class Cache extends Getters
      * Initialization that sets a base key and the driver based on configuration settings
      *
      * @param  Grav $grav
-     *
      * @return void
      */
     public function init(Grav $grav)
     {
-        /** @var Config $config */
         $this->config = $grav['config'];
         $this->now = time();
 
@@ -135,7 +141,7 @@ class Cache extends Getters
         $uniqueness = substr(md5($uri->rootUrl(true) . $this->config->key() . GRAV_VERSION), 2, 8);
 
         // Cache key allows us to invalidate all cache on configuration changes.
-        $this->key = ($prefix ? $prefix : 'g') . '-' . $uniqueness;
+        $this->key = ($prefix ?: 'g') . '-' . $uniqueness;
         $this->cache_dir = $grav['locator']->findResource('cache://doctrine/' . $uniqueness, true, true);
         $this->driver_setting = $this->config->get('system.cache.driver');
         $this->driver = $this->getCacheDriver();
@@ -164,24 +170,75 @@ class Cache extends Getters
     }
 
     /**
-     * Deletes the old out of date file-based caches
+     * Deletes old cache files based on age
      *
      * @return int
      */
     public function purgeOldCache()
     {
-        $cache_dir = dirname($this->cache_dir);
-        $current = basename($this->cache_dir);
+        // Get the max age for cache files from config (default 30 days)
+        $max_age_days = $this->config->get('system.cache.purge_max_age_days', 30);
+        $max_age_seconds = $max_age_days * 86400; // Convert days to seconds
+        $now = time();
         $count = 0;
-
-        foreach (new \DirectoryIterator($cache_dir) as $file) {
+        
+        // First, clean up old orphaned cache directories (not the current one)
+        $cache_dir = dirname($this->cache_dir);
+        $current = Utils::basename($this->cache_dir);
+        
+        foreach (new DirectoryIterator($cache_dir) as $file) {
             $dir = $file->getBasename();
             if ($dir === $current || $file->isDot() || $file->isFile()) {
                 continue;
             }
-
-            Folder::delete($file->getPathname());
-            $count++;
+            
+            // Check if directory is old and empty or very old (90+ days)
+            $dir_age = $now - $file->getMTime();
+            if ($dir_age > 7776000) { // 90 days
+                Folder::delete($file->getPathname());
+                $count++;
+            }
+        }
+        
+        // Now clean up old cache files within the current cache directory
+        if (is_dir($this->cache_dir)) {
+            $iterator = new \RecursiveIteratorIterator(
+                new \RecursiveDirectoryIterator($this->cache_dir, \RecursiveDirectoryIterator::SKIP_DOTS),
+                \RecursiveIteratorIterator::CHILD_FIRST
+            );
+            
+            foreach ($iterator as $file) {
+                if ($file->isFile()) {
+                    $file_age = $now - $file->getMTime();
+                    if ($file_age > $max_age_seconds) {
+                        @unlink($file->getPathname());
+                        $count++;
+                    }
+                }
+            }
+        }
+        
+        // Also clean up old files in compiled cache
+        $grav = Grav::instance();
+        $compiled_dir = $this->config->get('system.cache.compiled_dir', 'cache://compiled');
+        $compiled_path = $grav['locator']->findResource($compiled_dir, true);
+        
+        if ($compiled_path && is_dir($compiled_path)) {
+            $iterator = new \RecursiveIteratorIterator(
+                new \RecursiveDirectoryIterator($compiled_path, \RecursiveDirectoryIterator::SKIP_DOTS),
+                \RecursiveIteratorIterator::CHILD_FIRST
+            );
+            
+            foreach ($iterator as $file) {
+                if ($file->isFile()) {
+                    $file_age = $now - $file->getMTime();
+                    // Compiled files can be kept longer (60 days)
+                    if ($file_age > ($max_age_seconds * 2)) {
+                        @unlink($file->getPathname());
+                        $count++;
+                    }
+                }
+            }
         }
 
         return $count;
@@ -191,6 +248,7 @@ class Cache extends Getters
      * Public accessor to set the enabled state of the cache
      *
      * @param bool|int $enabled
+     * @return void
      */
     public function setEnabled($enabled)
     {
@@ -260,24 +318,28 @@ class Cache extends Getters
             case 'memcache':
                 if (extension_loaded('memcache')) {
                     $memcache = new \Memcache();
-                    $memcache->connect($this->config->get('system.cache.memcache.server', 'localhost'),
-                        $this->config->get('system.cache.memcache.port', 11211));
+                    $memcache->connect(
+                        $this->config->get('system.cache.memcache.server', 'localhost'),
+                        $this->config->get('system.cache.memcache.port', 11211)
+                    );
                     $driver = new DoctrineCache\MemcacheCache();
                     $driver->setMemcache($memcache);
                 } else {
-                    throw new \LogicException('Memcache PHP extension has not been installed');
+                    throw new LogicException('Memcache PHP extension has not been installed');
                 }
                 break;
 
             case 'memcached':
                 if (extension_loaded('memcached')) {
                     $memcached = new \Memcached();
-                    $memcached->addServer($this->config->get('system.cache.memcached.server', 'localhost'),
-                        $this->config->get('system.cache.memcached.port', 11211));
+                    $memcached->addServer(
+                        $this->config->get('system.cache.memcached.server', 'localhost'),
+                        $this->config->get('system.cache.memcached.port', 11211)
+                    );
                     $driver = new DoctrineCache\MemcachedCache();
                     $driver->setMemcached($memcached);
                 } else {
-                    throw new \LogicException('Memcached PHP extension has not been installed');
+                    throw new LogicException('Memcached PHP extension has not been installed');
                 }
                 break;
 
@@ -286,12 +348,15 @@ class Cache extends Getters
                     $redis = new \Redis();
                     $socket = $this->config->get('system.cache.redis.socket', false);
                     $password = $this->config->get('system.cache.redis.password', false);
+                    $databaseId = $this->config->get('system.cache.redis.database', 0);
 
                     if ($socket) {
                         $redis->connect($socket);
                     } else {
-                        $redis->connect($this->config->get('system.cache.redis.server', 'localhost'),
-                            $this->config->get('system.cache.redis.port', 6379));
+                        $redis->connect(
+                            $this->config->get('system.cache.redis.server', 'localhost'),
+                            $this->config->get('system.cache.redis.port', 6379)
+                        );
                     }
 
                     // Authenticate with password if set
@@ -299,10 +364,15 @@ class Cache extends Getters
                         throw new \RedisException('Redis authentication failed');
                     }
 
+                    // Select alternate ( !=0 ) database ID if set
+                    if ($databaseId && !$redis->select($databaseId)) {
+                        throw new \RedisException('Could not select alternate Redis database ID');
+                    }
+
                     $driver = new DoctrineCache\RedisCache();
                     $driver->setRedis($redis);
                 } else {
-                    throw new \LogicException('Redis PHP extension has not been installed');
+                    throw new LogicException('Redis PHP extension has not been installed');
                 }
                 break;
 
@@ -318,8 +388,7 @@ class Cache extends Getters
      * Gets a cached entry if it exists based on an id. If it does not exist, it returns false
      *
      * @param  string $id the id of the cached entry
-     *
-     * @return object|bool     returns the cached entry, can be any type, or false if doesn't exist
+     * @return mixed|bool     returns the cached entry, can be any type, or false if doesn't exist
      */
     public function fetch($id)
     {
@@ -334,8 +403,8 @@ class Cache extends Getters
      * Stores a new cached entry.
      *
      * @param  string       $id       the id of the cached entry
-     * @param  array|object $data     the data for the cached entry to store
-     * @param  int          $lifetime the lifetime to store the entry in seconds
+     * @param  array|object|int $data     the data for the cached entry to store
+     * @param  int|null     $lifetime the lifetime to store the entry in seconds
      */
     public function save($id, $data, $lifetime = null)
     {
@@ -393,6 +462,8 @@ class Cache extends Getters
 
     /**
      * Getter method to get the cache key
+     *
+     * @return string
      */
     public function getKey()
     {
@@ -401,6 +472,9 @@ class Cache extends Getters
 
     /**
      * Setter method to set key (Advanced)
+     *
+     * @param string $key
+     * @return void
      */
     public function setKey($key)
     {
@@ -412,7 +486,6 @@ class Cache extends Getters
      * Helper method to clear all Grav caches
      *
      * @param string $remove standard|all|assets-only|images-only|cache-only
-     *
      * @return array
      */
     public static function clearCache($remove = 'standard')
@@ -446,7 +519,6 @@ class Cache extends Getters
                 } else {
                     $remove_paths = self::$standard_remove_no_images;
                 }
-
         }
 
         // Delete entries in the doctrine cache if required
@@ -459,11 +531,12 @@ class Cache extends Getters
         Grav::instance()->fireEvent('onBeforeCacheClear', new Event(['remove' => $remove, 'paths' => &$remove_paths]));
 
         foreach ($remove_paths as $stream) {
-
             // Convert stream to a real path
             try {
                 $path = $locator->findResource($stream, true, true);
-                if($path === false) continue;
+                if ($path === false) {
+                    continue;
+                }
 
                 $anything = false;
                 $files = glob($path . '/*');
@@ -477,7 +550,7 @@ class Cache extends Getters
                                 $anything = true;
                             }
                         } elseif (is_dir($file)) {
-                            if (Folder::delete($file)) {
+                            if (Folder::delete($file, false)) {
                                 $anything = true;
                             }
                         }
@@ -487,7 +560,7 @@ class Cache extends Getters
                 if ($anything) {
                     $output[] = '<red>Cleared:  </red>' . $path . '/*';
                 }
-            } catch (\Exception $e) {
+            } catch (Exception $e) {
                 // stream not found or another error while deleting files.
                 $output[] = '<red>ERROR: </red>' . $e->getMessage();
             }
@@ -510,9 +583,14 @@ class Cache extends Getters
             @opcache_reset();
         }
 
+        Grav::instance()->fireEvent('onAfterCacheClear', new Event(['remove' => $remove, 'output' => &$output]));
+
         return $output;
     }
 
+    /**
+     * @return void
+     */
     public static function invalidateCache()
     {
         $user_config = USER_DIR . 'config/system.yaml';
@@ -528,13 +606,13 @@ class Cache extends Getters
         if (function_exists('opcache_reset')) {
             @opcache_reset();
         }
-
     }
 
     /**
      * Set the cache lifetime programmatically
      *
      * @param int $future timestamp
+     * @return void
      */
     public function setLifetime($future)
     {
@@ -552,7 +630,7 @@ class Cache extends Getters
     /**
      * Retrieve the cache lifetime (in seconds)
      *
-     * @return mixed
+     * @return int
      */
     public function getLifetime()
     {
@@ -566,7 +644,7 @@ class Cache extends Getters
     /**
      * Returns the current driver name
      *
-     * @return mixed
+     * @return string
      */
     public function getDriverName()
     {
@@ -576,7 +654,7 @@ class Cache extends Getters
     /**
      * Returns the current driver setting
      *
-     * @return mixed
+     * @return string
      */
     public function getDriverSetting()
     {
@@ -591,29 +669,37 @@ class Cache extends Getters
      */
     public function isVolatileDriver($setting)
     {
-        if (in_array($setting, ['apc', 'apcu', 'xcache', 'wincache'])) {
-            return true;
-        }
-
-        return false;
+        return in_array($setting, ['apc', 'apcu', 'xcache', 'wincache'], true);
     }
 
     /**
      * Static function to call as a scheduled Job to purge old Doctrine files
+     *
+     * @param bool $echo
+     *
+     * @return string|void
      */
-    public static function purgeJob()
+    public static function purgeJob($echo = false)
     {
         /** @var Cache $cache */
         $cache = Grav::instance()['cache'];
-        $deleted_folders = $cache->purgeOldCache();
+        $deleted_items = $cache->purgeOldCache();
+        
+        $max_age = $cache->config->get('system.cache.purge_max_age_days', 30);
+        $msg = 'Purged ' . $deleted_items . ' old cache items (files older than ' . $max_age . ' days)';
 
-        echo 'Purged ' . $deleted_folders . ' old cache folders...';
+        if ($echo) {
+            echo $msg;
+        } else {
+            return $msg;
+        }
     }
 
     /**
      * Static function to call as a scheduled Job to clear Grav cache
      *
      * @param string $type
+     * @return void
      */
     public static function clearJob($type)
     {
@@ -623,6 +709,10 @@ class Cache extends Getters
         echo strip_tags(implode("\n", $result));
     }
 
+    /**
+     * @param Event $event
+     * @return void
+     */
     public function onSchedulerInitialized(Event $event)
     {
         /** @var Scheduler $scheduler */
@@ -634,7 +724,7 @@ class Cache extends Getters
         $name = 'cache-purge';
         $logs = 'logs/' . $name . '.out';
 
-        $job = $scheduler->addFunction('Grav\Common\Cache::purgeJob', [], $name );
+        $job = $scheduler->addFunction('Grav\Common\Cache::purgeJob', [true], $name);
         $job->at($at);
         $job->output($logs);
         $job->backlink('/config/system#caching');
@@ -645,12 +735,9 @@ class Cache extends Getters
         $name = 'cache-clear';
         $logs = 'logs/' . $name . '.out';
 
-        $job = $scheduler->addFunction('Grav\Common\Cache::clearJob', [$clear_type], $name );
+        $job = $scheduler->addFunction('Grav\Common\Cache::clearJob', [$clear_type], $name);
         $job->at($at);
         $job->output($logs);
         $job->backlink('/config/system#caching');
-
     }
-
-
 }
